+++ 02-ignite-timer/src/pages/Home/components/NewCycleForm/index.tsx
@@ -1,0 +1,61 @@
+import { useForm } from "react-hook-form";
+
+import { zodResolver } from "@hookform/resolvers/zod";
+import * as zod from "zod";
+
+import { FormContainer, MinutesAmountInput, TaskInput } from "./styles";
+
+const newCycleFormValidationSchema = zod.object({
+  task: zod.string().min(1, "Informe a Tarefa"),
+  minutesAmount: zod
+    .number()
+    .min(1, "O cíclo precisa ser de no mínimo 5 minutos.")
+    .max(60, "O cíclo precisa ser de no máximo 60 minutos."),
+});
+
+type NewCycleFormData = zod.infer<typeof newCycleFormValidationSchema>;
+
+export function NewCycleForm() {
+  const { register, handleSubmit, watch, reset } = useForm<NewCycleFormData>({
+    resolver: zodResolver(newCycleFormValidationSchema),
+    defaultValues: {
+      task: "",
+      minutesAmount: 5,
+    },
+  });
+
+  return (
+    <FormContainer>
+      <label htmlFor="task">Vou trabalhar em</label>
+      <TaskInput
+        type="text"
+        id="task"
+        placeholder="Dê um nome para o seu projeto"
+        list="task-suggestion"
+        disabled={!!activeCycle}
+        {...register("task")}
+      />
+
+      <datalist id="task-suggestion">
+        <option value="Projeto 1" />
+        <option value="Projeto 2" />
+        <option value="Projeto 3" />
+        <option value="Projeto 4" />
+      </datalist>
+
+      <label htmlFor="minutesAmount">durante</label>
+      <MinutesAmountInput
+        type="number"
+        id="minutesAmount"
+        placeholder="00"
+        step={5}
+        min={1}
+        max={60}
+        disabled={!!activeCycle}
+        {...register("minutesAmount", { valueAsNumber: true })}
+      />
+
+      <span>minutos.</span>
+    </FormContainer>
+  );
+}
